requirements.txt
pytz
chardet
celery
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
pillow
prompt_toolkit
psycopg2-binary
PyJWT
python-dateutil
python-dotenv
requests
six
sqlparse
tzdata
urllib3





